ui additions and changes
- aesthetic changes to start ui
- added saves button to start menu. no functionality
- created saves canvas and added Saves to Menu enum. currently is just a copy of the settings canvas
- created SavesUI

large summary of this commit is ui
changes with basic logic implemented (like buttons). also the goback method might be broken but not completely confirmed yet. would have to have a case where the menu before the current one gets turned OFF. this would possibly cause problems because the goback method doesnt reopen the menu. a simple check would be if the menu to open (the one before the current one) is not on, turn it on, bypassing the use of OpenMenu